20.6.16

+++ pathicon/script.js
@@ -6,21 +6,17 @@ Addons.PathIcon = {
 	fStyle: LVIS_CUT | LVIS_SELECTED,
 	CONFIG: fso.BuildPath(te.Data.DataFolder, "config\\pathicon.tsv"),
 
-	GetIconImage: function (fn, Large)
-	{
+	GetIconImage: function (fn, Large) {
 		fn = api.PathUnquoteSpaces(ExtractMacro(te, fn));
 		return api.CreateObject("WICBitmap").FromFile(fn) || MakeImgData(fn, 0, Large ? 48 : 16);
 	},
 
-	Exec: function (Ctrl, pt)
-	{
-		AddonOptions("pathicon", function ()
-		{
+	Exec: function (Ctrl, pt) {
+		AddonOptions("pathicon", function () {
 		}, { FV: GetFolderView(Ctrl, pt) });
 	},
 
-	Replace: function (path, s, l)
-	{
+	Replace: function (path, s, l) {
 		path = path.toLowerCase();
 		var db = Addons.PathIcon.Icon[path];
 		if (!db) {
@@ -39,8 +35,7 @@ Addons.PathIcon = {
 		api.RedrawWindow(te.hwnd, null, 0, RDW_INVALIDATE | RDW_FRAME | RDW_ALLCHILDREN);
 	},
 
-	Remove: function (path, mode)
-	{
+	Remove: function (path, mode) {
 		path = path.toLowerCase();
 		var db = Addons.PathIcon.Icon[path];
 		if (db) {
@@ -62,8 +57,7 @@ Addons.PathIcon = {
 		}
 	},
 
-	ENumCB: function (fncb)
-	{
+	ENumCB: function (fncb) {
 		for (var path in Addons.PathIcon.Icon) {
 			var db = Addons.PathIcon.Icon[path];
 			fncb(path, db[0], db[1]);
@@ -72,8 +66,7 @@ Addons.PathIcon = {
 };
 
 if (window.Addon == 1) {
-	AddEvent("HandleIcon", function (Ctrl, pid)
-	{
+	AddEvent("HandleIcon", function (Ctrl, pid, iItem) {
 		if (Ctrl.hwndList && pid) {
 			var i = Ctrl.IconSize < 32 ? 0 : 1, db = Addons.PathIcon.Icon[api.GetDisplayNameOf(pid, SHGDN_FORADDRESSBAR | SHGDN_FORPARSING | SHGDN_ORIGINAL).toLowerCase()];
 			if (db) {
@@ -91,8 +84,7 @@ if (window.Addon == 1) {
 		}
 	}, true);
 
-	AddEvent("ItemPostPaint", function (Ctrl, pid, nmcd, vcd)
-	{
+	AddEvent("ItemPostPaint", function (Ctrl, pid, nmcd, vcd) {
 		var db = Addons.PathIcon.Icon[api.GetDisplayNameOf(pid, SHGDN_FORADDRESSBAR | SHGDN_FORPARSING | SHGDN_ORIGINAL).toLowerCase()];
 		if (db) {
 			var hList = Ctrl.hwndList;
@@ -128,60 +120,58 @@ if (window.Addon == 1) {
 				}
 				if (/object/i.test(typeof image)) {
 					var rc = api.Memory("RECT");
-					rc.Left = nmcd.dwItemSpec;
-					api.SendMessage(hTree, TVM_GETITEMRECT, true, rc);
-					var x = rc.Left - cx - 3 * screen.logicalYDPI / 96;
-					var w = image.GetWidth();
-					var h = image.GetHeight();
-					var y = rc.Top + (rc.Bottom - rc.Top - h) / 2;
-					var rc = api.Memory("RECT");
-					rc.left = x;
-					rc.right = x + w;
-					var n = h;
-					var cl = api.GetPixel(nmcd.hdc, x + w , y + h);
-					while (h-- > 0) {
-						var cl1 = api.GetPixel(nmcd.hdc, x + w , y + h);
-						if (cl == cl1 && h) {
-							continue;
+					rc.Write(0, VT_I8, nmcd.dwItemSpec);
+					if (api.SendMessage(hTree, TVM_GETITEMRECT, true, rc)) {
+						var x = rc.Left - cx - 3 * screen.logicalYDPI / 96;
+						var w = image.GetWidth();
+						var h = image.GetHeight();
+						var y = rc.Top + (rc.Bottom - rc.Top - h) / 2;
+						var rc = api.Memory("RECT");
+						rc.left = x;
+						rc.right = x + w;
+						var n = h;
+						var cl = api.GetPixel(nmcd.hdc, x + w, y + h);
+						while (h-- > 0) {
+							var cl1 = api.GetPixel(nmcd.hdc, x + w, y + h);
+							if (cl == cl1 && h) {
+								continue;
+							}
+							var brush = api.CreateSolidBrush(cl1);
+							rc.top = y + h;
+							rc.bottom = y + n;
+							api.FillRect(nmcd.hdc, rc, brush);
+							api.DeleteObject(brush);
+							cl = cl1;
+							n = h;
 						}
-						var brush = api.CreateSolidBrush(cl1);
-						rc.top = y + h;
-						rc.bottom = y + n;
-						api.FillRect(nmcd.hdc, rc, brush);
-						api.DeleteObject(brush);
-						cl = cl1;
-						n = h;
+						image.DrawEx(nmcd.hdc, x, y, 0, 0, CLR_NONE, CLR_NONE, ILD_NORMAL);
+						return S_OK;
 					}
-					image.DrawEx(nmcd.hdc, x, y, 0, 0, CLR_NONE, CLR_NONE, ILD_NORMAL);
-					return S_OK;
 				}
 			}
 		}
 	}, true);
 
-	AddEvent("GetIconImage", function (Ctrl, BGColor)
-	{
+	AddEvent("GetIconImage", function (Ctrl, BGColor) {
 		var db = Addons.PathIcon.Icon[(api.GetDisplayNameOf(Ctrl, SHGDN_FORADDRESSBAR | SHGDN_FORPARSING | SHGDN_ORIGINAL) || "").toLowerCase()];
 		if (db && db[0]) {
 			return MakeImgSrc(db[0], 0, false, 16);
 		}
 	});
 
-	AddEvent("SaveConfig", function ()
-	{
+	AddEvent("SaveConfig", function () {
 		if (Addons.PathIcon.bSave) {
 			try {
 				var ado = api.CreateObject("ads");
 				ado.CharSet = "utf-8";
 				ado.Open();
-				Addons.PathIcon.ENumCB(function (path, s, l)
-				{
+				Addons.PathIcon.ENumCB(function (path, s, l) {
 					ado.WriteText([path, s, l].join("\t") + "\r\n");
 				});
 				ado.SaveToFile(Addons.PathIcon.CONFIG, adSaveCreateOverWrite);
 				ado.Close();
 				Addons.PathIcon.bSave = false;
-			} catch (e) {}
+			} catch (e) { }
 		}
 	});
 
@@ -189,8 +179,7 @@ if (window.Addon == 1) {
 	//Menu
 	if (item.getAttribute("MenuExec")) {
 		Addons.PathIcon.nPos = api.LowPart(item.getAttribute("MenuPos"));
-		AddEvent(item.getAttribute("Menu"), function (Ctrl, hMenu, nPos)
-		{
+		AddEvent(item.getAttribute("Menu"), function (Ctrl, hMenu, nPos) {
 			api.InsertMenu(hMenu, Addons.PathIcon.nPos, MF_BYPOSITION | MF_STRING, ++nPos, Addons.PathIcon.strName);
 			ExtraMenuCommand[nPos] = Addons.PathIcon.Exec;
 			return nPos;
@@ -232,11 +221,10 @@ if (window.Addon == 1) {
 			}
 			ado.Close();
 		}
-	} catch (e) {}
+	} catch (e) { }
 
 	if (api.IsAppThemed() && WINVER >= 0x600) {
-		AddEvent("Load", function ()
-		{
+		AddEvent("Load", function () {
 			if (!Addons.ClassicStyle) {
 				Addons.PathIcon.fStyle = LVIS_CUT;
 			}
